Initial version of rqt_current_limit_controller
The logic is very similar to that of rqt_joint_trajectory_controller.
I managed to refactor some common blocks, but there is more I'd like to
make shared.

For the record, the dynamic_reconfigure approach was not taken because
in a rqt plugin we have more control over the widgets, which allows to
implement the monitor/command modes.